Deployed a new subgraph
