--- server/subgraph/dbeats/networks.json
@@ -1,8 +1,0 @@
-{
-  "arbitrum-sepolia": {
-    "DBeatsFactory": {
-      "address": "0x2242cF684DAC41CC45a8B07eF5E027dc28322e49",
-      "startBlock": 68281770
-    }
-  }
-}
