--- server/subgraph/dbeats/package.json
@@ -1,18 +1,0 @@
-{
-  "name": "dbeats",
-  "license": "UNLICENSED",
-  "scripts": {
-    "codegen": "graph codegen",
-    "build": "graph build",
-    "deploy": "graph deploy --node https://api.studio.thegraph.com/deploy/ dbeats",
-    "create-local": "graph create --node http://localhost:8020/ dbeats",
-    "remove-local": "graph remove --node http://localhost:8020/ dbeats",
-    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 dbeats",
-    "test": "graph test"
-  },
-  "dependencies": {
-    "@graphprotocol/graph-cli": "0.80.0",
-    "@graphprotocol/graph-ts": "0.32.0"
-  },
-  "devDependencies": { "matchstick-as": "0.5.0" }
-}
