--- server/subgraph/dbeats/schema.graphql
@@ -1,34 +1,0 @@
-type NFT @entity {
-  id: ID!
-  address: Bytes!
-  artist: Artist!
-  tokenURI: String!
-  name: String!
-  symbol: String!
-  mintPrice: BigInt!
-  genre: String!
-  createdAt: BigInt!
-}
-
-type Artist @entity {
-  id: ID!
-  address: Bytes!
-  nfts: [NFT!]! @derivedFrom(field: "artist")
-  createdAt: BigInt!
-}
-
-type Factory @entity {
-  id: ID!
-  nftsCount: BigInt!
-  artistsCount: BigInt!
-}
-
-type RoleGranted @entity {
-  id: ID!
-  role: Bytes!
-  account: Bytes!
-  sender: Bytes!
-  blockNumber: BigInt!
-  blockTimestamp: BigInt!
-  transactionHash: Bytes!
-}
