--- server/subgraph/dbeats/subgraph.yaml
@@ -1,29 +1,0 @@
-specVersion: 0.0.5
-schema:
-  file: ./schema.graphql
-dataSources:
-  - kind: ethereum
-    name: DBeatsFactory
-    network: arbitrum-sepolia
-    source:
-      address: "0x2242cF684DAC41CC45a8B07eF5E027dc28322e49"
-      abi: DBeatsFactory
-      startBlock: 68281770
-    mapping:
-      kind: ethereum/events
-      apiVersion: 0.0.7
-      language: wasm/assemblyscript
-      entities:
-        - NFT
-        - Artist
-        - Factory
-        - RoleGranted
-      abis:
-        - name: DBeatsFactory
-          file: ./abis/DBeatsFactory.json
-      eventHandlers:
-        - event: NewNFT(indexed address,address,string,string,string,uint256,string)
-          handler: handleNewNFT
-        - event: RoleGranted(indexed bytes32,indexed address,indexed address)
-          handler: handleRoleGranted
-      file: ./src/d-beats-factory.ts
