--- server/subgraph/dbeats/tests/d-beats-factory-utils.ts
@@ -1,154 +1,0 @@
-import { newMockEvent } from "matchstick-as"
-import { ethereum, Address, BigInt, Bytes } from "@graphprotocol/graph-ts"
-import {
-  NewNFT,
-  OwnershipTransferred,
-  RoleAdminChanged,
-  RoleGranted,
-  RoleRevoked
-} from "../generated/DBeatsFactory/DBeatsFactory"
-
-export function createNewNFTEvent(
-  nftAddress: Address,
-  _artistAddress: Address,
-  _newTokenURI: string,
-  name: string,
-  symbol: string,
-  mintPrice: BigInt,
-  _genre: string
-): NewNFT {
-  let newNftEvent = changetype<NewNFT>(newMockEvent())
-
-  newNftEvent.parameters = new Array()
-
-  newNftEvent.parameters.push(
-    new ethereum.EventParam(
-      "nftAddress",
-      ethereum.Value.fromAddress(nftAddress)
-    )
-  )
-  newNftEvent.parameters.push(
-    new ethereum.EventParam(
-      "_artistAddress",
-      ethereum.Value.fromAddress(_artistAddress)
-    )
-  )
-  newNftEvent.parameters.push(
-    new ethereum.EventParam(
-      "_newTokenURI",
-      ethereum.Value.fromString(_newTokenURI)
-    )
-  )
-  newNftEvent.parameters.push(
-    new ethereum.EventParam("name", ethereum.Value.fromString(name))
-  )
-  newNftEvent.parameters.push(
-    new ethereum.EventParam("symbol", ethereum.Value.fromString(symbol))
-  )
-  newNftEvent.parameters.push(
-    new ethereum.EventParam(
-      "mintPrice",
-      ethereum.Value.fromUnsignedBigInt(mintPrice)
-    )
-  )
-  newNftEvent.parameters.push(
-    new ethereum.EventParam("_genre", ethereum.Value.fromString(_genre))
-  )
-
-  return newNftEvent
-}
-
-export function createOwnershipTransferredEvent(
-  previousOwner: Address,
-  newOwner: Address
-): OwnershipTransferred {
-  let ownershipTransferredEvent = changetype<OwnershipTransferred>(
-    newMockEvent()
-  )
-
-  ownershipTransferredEvent.parameters = new Array()
-
-  ownershipTransferredEvent.parameters.push(
-    new ethereum.EventParam(
-      "previousOwner",
-      ethereum.Value.fromAddress(previousOwner)
-    )
-  )
-  ownershipTransferredEvent.parameters.push(
-    new ethereum.EventParam("newOwner", ethereum.Value.fromAddress(newOwner))
-  )
-
-  return ownershipTransferredEvent
-}
-
-export function createRoleAdminChangedEvent(
-  role: Bytes,
-  previousAdminRole: Bytes,
-  newAdminRole: Bytes
-): RoleAdminChanged {
-  let roleAdminChangedEvent = changetype<RoleAdminChanged>(newMockEvent())
-
-  roleAdminChangedEvent.parameters = new Array()
-
-  roleAdminChangedEvent.parameters.push(
-    new ethereum.EventParam("role", ethereum.Value.fromFixedBytes(role))
-  )
-  roleAdminChangedEvent.parameters.push(
-    new ethereum.EventParam(
-      "previousAdminRole",
-      ethereum.Value.fromFixedBytes(previousAdminRole)
-    )
-  )
-  roleAdminChangedEvent.parameters.push(
-    new ethereum.EventParam(
-      "newAdminRole",
-      ethereum.Value.fromFixedBytes(newAdminRole)
-    )
-  )
-
-  return roleAdminChangedEvent
-}
-
-export function createRoleGrantedEvent(
-  role: Bytes,
-  account: Address,
-  sender: Address
-): RoleGranted {
-  let roleGrantedEvent = changetype<RoleGranted>(newMockEvent())
-
-  roleGrantedEvent.parameters = new Array()
-
-  roleGrantedEvent.parameters.push(
-    new ethereum.EventParam("role", ethereum.Value.fromFixedBytes(role))
-  )
-  roleGrantedEvent.parameters.push(
-    new ethereum.EventParam("account", ethereum.Value.fromAddress(account))
-  )
-  roleGrantedEvent.parameters.push(
-    new ethereum.EventParam("sender", ethereum.Value.fromAddress(sender))
-  )
-
-  return roleGrantedEvent
-}
-
-export function createRoleRevokedEvent(
-  role: Bytes,
-  account: Address,
-  sender: Address
-): RoleRevoked {
-  let roleRevokedEvent = changetype<RoleRevoked>(newMockEvent())
-
-  roleRevokedEvent.parameters = new Array()
-
-  roleRevokedEvent.parameters.push(
-    new ethereum.EventParam("role", ethereum.Value.fromFixedBytes(role))
-  )
-  roleRevokedEvent.parameters.push(
-    new ethereum.EventParam("account", ethereum.Value.fromAddress(account))
-  )
-  roleRevokedEvent.parameters.push(
-    new ethereum.EventParam("sender", ethereum.Value.fromAddress(sender))
-  )
-
-  return roleRevokedEvent
-}
